requirements.txt
Flask==2.1.0
openai
python-dotenv
requests==2.26.0
whoosh
beautifulsoup4==4.10.0
lxml==4.6.3